requirements.txt

Flask==3.0.0
Jinja2==3.1.2
mysql-connector-python
